Unit tests for 502 status code + disabled auto send, long => uint api change, adjusted metrics initialization flow

+++ Runtime/Interfaces/IBacktraceClient.cs
@@ -65,6 +65,6 @@ namespace Backtrace.Unity.Interfaces
         /// <summary>
         /// Enable event aggregation support.
         /// </summary>
-        void EnableMetrics(string uniqueEventsSubmissionUrl, string summedEventsSubmissionUrl, long timeIntervalInSec, string uniqueEventName);
+        void EnableMetrics(string uniqueEventsSubmissionUrl, string summedEventsSubmissionUrl, uint timeIntervalInSec, string uniqueEventName);
     }
 }
